Couldn't click next line button, when table is resized. #831 Issue: https://github.com/xdan/jodit/issues/831
