+++ src/plugins/add-new-line/add-new-line.ts
@@ -13,7 +13,6 @@
 import './add-new-line.less';
 
 import type { IBound, IJodit, HTMLTagNames, Nullable } from 'jodit/types';
-import { Config } from 'jodit/config';
 import { Dom, Icon, Plugin } from 'jodit/modules';
 import {
 	offset,
@@ -21,49 +20,9 @@ import {
 	call,
 	scrollIntoViewIfNeeded
 } from 'jodit/core/helpers';
+import { debounce, watch } from 'jodit/core/decorators';
 
-declare module 'jodit/config' {
-	interface Config {
-		/**
-		 * Create helper
-		 */
-		addNewLine: boolean;
-
-		/**
-		 * What kind of tags it will be impact
-		 */
-		addNewLineTagsTriggers: HTMLTagNames[];
-
-		/**
-		 * On dbl click on empty space of editor it add new P element
-		 * @example
-		 * ```js
-		 * Jodit.make('#editor', {
-		 *   addNewLineOnDBLClick: false // disable
-		 * })
-		 * ```
-		 */
-		addNewLineOnDBLClick: boolean;
-
-		/**
-		 * Absolute delta between cursor position and edge(top or bottom)
-		 * of element when show line
-		 */
-		addNewLineDeltaShow: number;
-	}
-}
-
-Config.prototype.addNewLine = true;
-Config.prototype.addNewLineOnDBLClick = true;
-Config.prototype.addNewLineTagsTriggers = [
-	'table',
-	'iframe',
-	'img',
-	'hr',
-	'pre',
-	'jodit'
-];
-Config.prototype.addNewLineDeltaShow = 20;
+import './config';
 
 const ns = 'addnewline';
 
@@ -113,6 +72,7 @@ export class addNewLine extends Plugin {
 		this.j.async.clearTimeout(this.timeout);
 		this.lineInFocus = false;
 		Dom.safeRemove(this.line);
+		this.line.style.setProperty('--jd-offset-handle', '0');
 	};
 
 	private hide = (): void => {
@@ -171,14 +131,7 @@ export class addNewLine extends Plugin {
 			.on(editor.editor, 'dblclick' + '.' + ns, this.onDblClickEditor)
 			.on(editor.editor, 'click' + '.' + ns, this.hide)
 			.on(editor.container, 'mouseleave' + '.' + ns, this.hide)
-			.on(
-				editor.editor,
-				'mousemove' + '.' + ns,
-				editor.async.debounce(
-					this.onMouseMove,
-					editor.defaultTimeout * 3
-				)
-			);
+			.on(editor.editor, 'mousemove' + '.' + ns, this.onMouseMove);
 	}
 
 	private onClickLine = (e: MouseEvent): void => {
@@ -186,7 +139,11 @@ export class addNewLine extends Plugin {
 		const p = editor.createInside.element(editor.o.enter);
 
 		if (this.preview && this.current && this.current.parentNode) {
-			this.current.parentNode.insertBefore(p, this.current);
+			if (this.current === editor.editor) {
+				Dom.prepend(editor.editor, p);
+			} else {
+				this.current.parentNode.insertBefore(p, this.current);
+			}
 		} else {
 			editor.editor.appendChild(p);
 		}
@@ -200,7 +157,8 @@ export class addNewLine extends Plugin {
 		e.preventDefault();
 	};
 
-	private onDblClickEditor = (e: MouseEvent): void => {
+	@watch(':dblclick')
+	private onDblClickEditor(e: MouseEvent): void {
 		const editor = this.j;
 
 		if (
@@ -235,9 +193,10 @@ export class addNewLine extends Plugin {
 			this.hideForce();
 			e.preventDefault();
 		}
-	};
+	}
 
-	private onMouseMove = (e: MouseEvent): void => {
+	@debounce()
+	private onMouseMove(e: MouseEvent): void {
 		const editor = this.j;
 
 		let currentElement: HTMLElement | null = editor.ed.elementFromPoint(
@@ -256,7 +215,10 @@ export class addNewLine extends Plugin {
 			return;
 		}
 
-		if (!this.isMatchedTag(currentElement)) {
+		if (
+			editor.editor !== currentElement &&
+			!this.isMatchedTag(currentElement)
+		) {
 			currentElement = Dom.closest(
 				currentElement,
 				this.isMatchedTag,
@@ -306,21 +268,26 @@ export class addNewLine extends Plugin {
 
 		if (
 			top !== false &&
-			!call(
-				this.preview ? Dom.prev : Dom.next,
-				currentElement,
-				this.canGetFocus,
-				editor.editor
-			)
+			((editor.editor === currentElement && !this.preview) ||
+				!call(
+					this.preview ? Dom.prev : Dom.next,
+					currentElement,
+					this.canGetFocus,
+					editor.editor
+				))
 		) {
 			this.line.style.top = top + 'px';
 			this.current = currentElement;
 			this.show();
+			this.line.style.setProperty(
+				'--jd-offset-handle',
+				e.clientX - pos.left - 10 + 'px'
+			);
 		} else {
 			this.current = false;
 			this.hide();
 		}
-	};
+	}
 
 	/** @override */
 	protected beforeDestruct(): void {
